寿命の計算をfloat型に変更 寿命に基づいた計算をFPS準拠に変更完了

+++ particleLifeImgui.h
@@ -8,7 +8,7 @@ class ParticleLifeImgui : public ImguiBase
 {
 private:
 	//パーティクルの変更可能ステータス : ライフ
-	int m_LifeSlider{};
+	float m_LifeSlider{};
 
 public:
 	using ImguiBase::ImguiBase;
